float to double

+++ src/com/ml/lib/autograd/AutoGrad.java
@@ -30,7 +30,7 @@ public class AutoGrad implements AutoGradInterface {
     public static Tensor neg(Tensor t1){
         return t1.getAutoGrad()._method_(new Neg());
     }
-    public static Tensor pow(Tensor tensor, float pow){
+    public static Tensor pow(Tensor tensor, double pow){
         return tensor.getAutoGrad()._method_(new Pow(pow));
     }
