+++ src/com/ml/lib/linear_algebra/operations/elementary/Sub.java
@@ -15,7 +15,7 @@ public class Sub extends ElementByElement {
     //-------------------------------------
 
     @Override
-    public float operation(float a, float b){
+    public double operation(double a, double b){
         return a - b;
     }
