+++ src/com/ml/lib/linear_algebra/operations/elementary/Sum.java
@@ -18,7 +18,7 @@ public class Sum extends ElementByElement {
     //-------------------------------------
 
     @Override
-    public float operation(float a, float b){
+    public double operation(double a, double b){
         return a + b;
     }
